homeassistant_salus improvements
