+++ salus_it600/gateway.py
@@ -13,38 +13,29 @@ from aiohttp import client_exceptions
 from .const import (
     COVER_POSITION_MAX,
     COVER_POSITION_MIN,
-    CURRENT_HVAC_HEAT,
-    CURRENT_HVAC_HEAT_IDLE,
-    CURRENT_HVAC_COOL,
-    CURRENT_HVAC_COOL_IDLE,
-    CURRENT_HVAC_IDLE,
-    CURRENT_HVAC_OFF,
     HVAC_MODE_HEAT,
     HVAC_MODE_COOL,
     HVAC_MODE_OFF,
-    HVAC_MODE_AUTO,
-    PRESET_FOLLOW_SCHEDULE,
     PRESET_OFF,
     PRESET_PERMANENT_HOLD,
     PRESET_TEMPORARY_HOLD,
     PRESET_ECO,
-    SUPPORT_FAN_MODE,
-    SUPPORT_PRESET_MODE,
-    SUPPORT_TARGET_TEMPERATURE,
-    TEMP_CELSIUS,
-    SUPPORT_OPEN,
-    SUPPORT_CLOSE,
-    SUPPORT_SET_POSITION,
     FAN_MODE_AUTO,
     FAN_MODE_HIGH,
     FAN_MODE_MEDIUM,
     FAN_MODE_LOW,
-    FAN_MODE_OFF,
     FanMode,
     HoldType,
-    RunningState,
     SystemMode,
     TEMPERATURE_SCALE,
+)
+from ._device_parsers import (
+    PARSING_EXCEPTIONS,
+    parse_binary_sensor_device,
+    parse_climate_device,
+    parse_cover_device,
+    parse_sensor_device,
+    parse_switch_device,
 )
 from .encryptor import IT600Encryptor
 from .exceptions import (
@@ -53,24 +44,15 @@ from .exceptions import (
     IT600ConnectionError,
 )
 from .device_models import (
-    BINARY_RELAY_MODELS,
     MODEL_FC600,
-    SKIPPED_BINARY_SENSOR_MODELS,
-    binary_sensor_device_class,
     is_binary_sensor_summary,
     is_sq610_model,
-    model_identifier,
-    switch_device_class,
 )
 from .models import GatewayDevice, ClimateDevice, BinarySensorDevice, SwitchDevice, CoverDevice, SensorDevice
 
 _LOGGER = logging.getLogger("salus_it600")
 
-DEFAULT_HOLD_TYPE = HoldType.PERMANENT_HOLD
-DEFAULT_RUNNING_STATE = RunningState.IDLE
 DEVICE_NOT_FOUND_ERROR = "{device_type} device not found with id: {device_id}"
-PARSING_EXCEPTIONS = (KeyError, TypeError, ValueError)
-_MISSING_HOLD_TYPE_WARNED: set[str] = set()
 DeviceT = TypeVar(
     "DeviceT",
     ClimateDevice,
@@ -168,34 +150,6 @@ async def _notify_update_callbacks(
         await update_callback(device_id=device_id)
 
 
-def _device_name(device_status: dict[str, Any], unique_id: str | None) -> str:
-    """Return a device name from a raw Salus gateway payload."""
-    default_name = unique_id or "Unknown"
-    raw_name = device_status.get("sZDO", {}).get(
-        "DeviceName",
-        json.dumps({"deviceName": default_name}),
-    )
-
-    try:
-        device_name = json.loads(raw_name)["deviceName"]
-        return device_name if isinstance(device_name, str) else default_name
-    except (KeyError, TypeError, ValueError):
-        return default_name
-
-
-def _hold_type(payload: dict[str, Any], unique_id: str) -> int:
-    """Return HoldType, defaulting broken payloads to permanent hold."""
-    if "HoldType" not in payload and unique_id not in _MISSING_HOLD_TYPE_WARNED:
-        _LOGGER.warning(
-            "Salus climate device %s is missing HoldType in the gateway payload; "
-            "treating it as Permanent Hold so the device can load",
-            unique_id,
-        )
-        _MISSING_HOLD_TYPE_WARNED.add(unique_id)
-
-    return int(payload.get("HoldType", DEFAULT_HOLD_TYPE))
-
-
 def _validate_gateway_response(response: Any, context: str) -> dict[str, Any]:
     """Validate that a gateway response is a JSON object with a status field."""
     if not isinstance(response, dict):
@@ -255,443 +209,22 @@ def _device_status_request_items(
     return request_items
 
 
-def _online(device_status: dict[str, Any]) -> bool:
-    """Return whether a device is marked online in a gateway payload."""
-    status = device_status.get("sZDOInfo", {}).get("OnlineStatus_i", 1)
-    return bool(status == 1)
+def _flatten_dict(data: dict[str, Any]) -> dict[str, Any]:
+    """Flatten nested gateway payload dictionaries into a single key/value map."""
+    flattened: dict[str, Any] = {}
 
+    def _walk(value: Any) -> None:
+        if not isinstance(value, dict):
+            return
 
-def _firmware_version(device_status: dict[str, Any]) -> str | None:
-    """Return the common firmware version field from a gateway payload."""
-    version = device_status.get("sZDO", {}).get("FirmwareVersion")
-    return version if isinstance(version, str) else None
+        for nested_key, nested_value in value.items():
+            if isinstance(nested_value, dict):
+                _walk(nested_value)
+            else:
+                flattened[nested_key] = nested_value
 
-
-def _common_device_args(
-    device_status: dict[str, Any],
-    unique_id: str,
-) -> dict[str, Any]:
-    """Return constructor args shared by most device models."""
-    model = model_identifier(device_status)
-    return {
-        "available": _online(device_status),
-        "name": _device_name(device_status, unique_id),
-        "unique_id": unique_id,
-        "data": device_status["data"],
-        "manufacturer": device_status.get("sBasicS", {}).get(
-            "ManufactureName",
-            "SALUS",
-        ),
-        "model": model,
-        "sw_version": _firmware_version(device_status),
-    }
-
-
-def _parse_cover_device(device_status: dict[str, Any]) -> CoverDevice | None:
-    """Parse one cover (roller shutter/blind) device from gateway payload.
-    
-    Extracts position, movement state, and capability info from cover-specific
-    protocol fields. Skips endpoints disabled via sButtonS.Mode.
-    
-    Protocol fields:
-    - `sLevelS.CurrentLevel`: Current position (0-100), 0=closed, 100=open
-    - `sLevelS.MoveToLevel_f`: Target position as hex string (first 2 chars)
-    - `sButtonS.Mode`: Endpoint enabled state (0=disabled)
-    
-    Args:
-        device_status: Cover device detail dict from gateway
-    
-    Returns:
-        CoverDevice model instance or None if:
-        - No UniID
-        - Endpoint disabled (Mode==0)
-    
-    Raises:
-        PARSING_EXCEPTIONS: If required fields missing/invalid (caught upstream)
-    """
-    unique_id = device_status.get("data", {}).get("UniID")
-    if unique_id is None:
-        return None
-
-    if device_status.get("sButtonS", {}).get("Mode") == 0:
-        return None
-
-    current_position = device_status.get("sLevelS", {}).get("CurrentLevel")
-    move_to_level = device_status.get("sLevelS", {}).get("MoveToLevel_f")
-    if move_to_level is not None and len(move_to_level) >= 2:
-        set_position = int(move_to_level[:2], 16)
-    else:
-        set_position = None
-
-    return CoverDevice(
-        **_common_device_args(device_status, unique_id),
-        current_cover_position=current_position,
-        is_opening=None if set_position is None else current_position < set_position,
-        is_closing=None if set_position is None else current_position > set_position,
-        is_closed=current_position == 0,
-        supported_features=SUPPORT_OPEN | SUPPORT_CLOSE | SUPPORT_SET_POSITION,
-        device_class=None,
-    )
-
-
-def _parse_switch_device(device_status: dict[str, Any]) -> SwitchDevice | None:
-    """Parse one switch (relay) device from gateway payload.
-    
-    Handles single and multi-endpoint switches. Multi-endpoint switches are
-    tracked with endpoint suffix in unique_id (e.g. "device-1_1" for endpoint 1).
-    Skips endpoints that are dimmers (have sLevelS).
-    
-    Protocol fields:
-    - `data.UniID`: Base device identifier
-    - `data.Endpoint`: Endpoint number (multi-endpoint devices)
-    - `sOnOffS.OnOff`: Relay state (0=off, 1=on)
-    - `sLevelS`: Present if dimmer (skip this endpoint)
-    
-    Args:
-        device_status: Switch device detail dict from gateway
-    
-    Returns:
-        SwitchDevice model instance or None if:
-        - No UniID
-        - Is dimmer (has sLevelS)
-        - OnOff field missing
-    
-    Raises:
-        PARSING_EXCEPTIONS: If required fields invalid (caught upstream)
-    """
-    base_unique_id = device_status.get("data", {}).get("UniID")
-    if base_unique_id is None:
-        return None
-
-    unique_id = f"{base_unique_id}_{device_status['data']['Endpoint']}"
-    if device_status.get("sLevelS") is not None:
-        return None
-
-    is_on = device_status.get("sOnOffS", {}).get("OnOff")
-    if is_on is None:
-        return None
-
-    model = model_identifier(device_status)
-    return SwitchDevice(
-        **_common_device_args(device_status, unique_id),
-        is_on=is_on == 1,
-        device_class=switch_device_class(model),
-    )
-
-
-def _parse_sensor_device(device_status: dict[str, Any]) -> SensorDevice | None:
-    """Parse one temperature sensor device from gateway payload.
-    
-    Extracts temperature measurement and sensor metadata. Some devices measure
-    temperature as secondary capability (e.g. SW600 window sensor with temp),
-    so unique_id is suffixed with "_temp" to avoid collision.
-    
-    Protocol fields:
-    - `sTempS.MeasuredValue_x100`: Temperature in 1/100ths of degree (e.g. 2150 = 21.5°C)
-    
-    Args:
-        device_status: Temperature sensor detail dict from gateway
-    
-    Returns:
-        SensorDevice model instance or None if:
-        - No UniID
-        - MeasuredValue_x100 field missing
-    
-    Raises:
-        PARSING_EXCEPTIONS: If required fields invalid (caught upstream)
-    """
-    unique_id = device_status.get("data", {}).get("UniID")
-    if unique_id is None:
-        return None
-
-    temperature = device_status.get("sTempS", {}).get("MeasuredValue_x100")
-    if temperature is None:
-        return None
-
-    unique_id = f"{unique_id}_temp"
-    return SensorDevice(
-        **_common_device_args(device_status, unique_id),
-        state=temperature / TEMPERATURE_SCALE,
-        unit_of_measurement=TEMP_CELSIUS,
-        device_class="temperature",
-    )
-
-
-def _parse_binary_sensor_device(
-    device_status: dict[str, Any],
-) -> BinarySensorDevice | None:
-    """Parse one binary sensor device from gateway payload.
-    
-    Handles two classes of binary sensors:
-    1. Standard IAS Zone Cluster (sIASZS): Contact/smoke/motion sensors
-    2. Binary relay models (MINITRV, Receiver): Single relay state
-    
-    Some models (SB600 button) are filtered out and return None.
-    
-    Protocol fields:
-    - `sIASZS.ErrorIASZSAlarmed1`: Alarm state for standard sensors (0/1)
-    - `sIT600I.RelayStatus`: Relay state for MINITRV/Receiver (0/1)
-    
-    Args:
-        device_status: Binary sensor detail dict from gateway
-    
-    Returns:
-        BinarySensorDevice model instance or None if:
-        - No UniID
-        - Model in SKIPPED_BINARY_SENSOR_MODELS (e.g. SB600)
-        - Required state field missing
-    
-    Raises:
-        PARSING_EXCEPTIONS: If required fields invalid (caught upstream)
-    """
-    unique_id = device_status.get("data", {}).get("UniID")
-    if unique_id is None:
-        return None
-
-    model = model_identifier(device_status)
-    if model in SKIPPED_BINARY_SENSOR_MODELS:
-        return None
-
-    if model in BINARY_RELAY_MODELS:
-        is_on = device_status.get("sIT600I", {}).get("RelayStatus")
-    else:
-        is_on = device_status.get("sIASZS", {}).get("ErrorIASZSAlarmed1")
-
-    if is_on is None:
-        return None
-
-    return BinarySensorDevice(
-        **_common_device_args(device_status, unique_id),
-        is_on=is_on == 1,
-        device_class=binary_sensor_device_class(model),
-    )
-
-
-def _climate_common_args(
-    device_status: dict[str, Any],
-    unique_id: str,
-) -> dict[str, Any]:
-    """Return constructor args shared by climate devices."""
-    return {
-        **_common_device_args(device_status, unique_id),
-        "temperature_unit": TEMP_CELSIUS,
-        "precision": 0.1,
-        "device_class": "temperature",
-    }
-
-
-def _parse_it600th_climate_device(
-    device_status: dict[str, Any],
-    unique_id: str,
-    th: dict[str, Any],
-) -> ClimateDevice:
-    """Parse one iT600 or SQ610 thermostat from gateway payload.
-    
-    Handles both standard iT600TH thermostats (heating-only) and SQ610 Quantum
-    thermostats (heating/cooling with extended controls).
-    
-    SQ610-specific behavior:
-    - Humidity read from SunnySetpoint_x100 field (divide by 100)
-    - Auto/Permanent/Standby hold modes (0/2/7)
-    - No fan mode support (always None)
-    
-    Standard iT600 behavior:
-    - Heat-only mode (no cooling)
-    - Simple hold modes: Off/Heat/Auto (7/2/0)
-    
-    Protocol fields:
-    - `sIT600TH.LocalTemperature_x100`: Current temperature (divide by 100)
-    - `sIT600TH.HeatingSetpoint_x100`: Target temperature (divide by 100)
-    - `sIT600TH.HoldType`: Preset mode (0=auto, 2=permanent, 7=off)
-    - `sIT600TH.RunningState`: Active state (even=idle, odd=heating)
-    - `sIT600TH.SunnySetpoint_x100`: Humidity for SQ610 (divide by 100)
-    
-    Args:
-        device_status: Full device detail dict from gateway
-        unique_id: Device UniID
-        th: The sIT600TH section of device_status
-    
-    Returns:
-        ClimateDevice model with appropriate mode/preset for this model variant
-    
-    Raises:
-        KeyError: If required sIT600TH fields missing (caught as PARSING_EXCEPTIONS)
-    """
-    model = model_identifier(device_status)
-    raw_humidity = th.get("SunnySetpoint_x100")
-    current_humidity = (
-        raw_humidity / TEMPERATURE_SCALE
-        if is_sq610_model(model) and isinstance(raw_humidity, int | float)
-        else None
-    )
-    hold_type = _hold_type(th, unique_id)
-    running_state = th.get("RunningState", DEFAULT_RUNNING_STATE)
-    return ClimateDevice(
-        **_climate_common_args(device_status, unique_id),
-        current_humidity=current_humidity,
-        current_temperature=th["LocalTemperature_x100"] / TEMPERATURE_SCALE,
-        target_temperature=th["HeatingSetpoint_x100"] / TEMPERATURE_SCALE,
-        max_temp=th.get("MaxHeatSetpoint_x100", 3500) / TEMPERATURE_SCALE,
-        min_temp=th.get("MinHeatSetpoint_x100", 500) / TEMPERATURE_SCALE,
-        hvac_mode=HVAC_MODE_OFF
-        if hold_type == HoldType.STANDBY
-        else HVAC_MODE_HEAT
-        if hold_type == HoldType.PERMANENT_HOLD
-        else HVAC_MODE_AUTO,
-        hvac_action=CURRENT_HVAC_OFF
-        if hold_type == HoldType.STANDBY
-        else CURRENT_HVAC_IDLE
-        if running_state % 2 == 0
-        else CURRENT_HVAC_HEAT,
-        hvac_modes=[HVAC_MODE_OFF, HVAC_MODE_HEAT, HVAC_MODE_AUTO],
-        preset_mode=PRESET_OFF
-        if hold_type == HoldType.STANDBY
-        else PRESET_PERMANENT_HOLD
-        if hold_type == HoldType.PERMANENT_HOLD
-        else PRESET_FOLLOW_SCHEDULE,
-        preset_modes=[PRESET_FOLLOW_SCHEDULE, PRESET_PERMANENT_HOLD, PRESET_OFF],
-        fan_mode=None,
-        fan_modes=None,
-        locked=None,
-        supported_features=SUPPORT_TARGET_TEMPERATURE | SUPPORT_PRESET_MODE,
-    )
-
-
-def _parse_fan_coil_climate_device(
-    device_status: dict[str, Any],
-    unique_id: str,
-    ther: dict[str, Any],
-    scomm: dict[str, Any],
-    sfans: dict[str, Any],
-) -> ClimateDevice:
-    """Parse one FC600 fan-coil thermostat from gateway payload.
-    
-    FC600 thermostats use a different protocol structure than iT600TH:
-    - Separate heating/cooling setpoints (sTherS.HeatingSetpoint_x100 and
-      sTherS.CoolingSetpoint_x100) depending on SystemMode
-    - Hold type in sComm section instead of sTherS
-    - Fan mode in sFanS section
-    - Support for heating, cooling, and auto modes
-    
-    Protocol fields:
-    - `sTherS.LocalTemperature_x100`: Current temperature (divide by 100)
-    - `sTherS.HeatingSetpoint_x100`: Target when heating (divide by 100)
-    - `sTherS.CoolingSetpoint_x100`: Target when cooling (divide by 100)
-    - `sTherS.SystemMode`: Mode (3=cool, 4=heat, else auto)
-    - `sComm.SetHoldType`: Preset mode (0=follow, 1=temp hold, 2=permanent, 7=off, 10=eco)
-    - `sFanS.FanMode`: Fan speed (0=off, 1=low, 2=medium, 3=high, 5=auto)
-    - `sTherS.RunningState`: Active state (0=idle, 33=heat, 66=cool)
-    
-    Args:
-        device_status: Full device detail dict from gateway
-        unique_id: Device UniID
-        ther: The sTherS section of device_status (with setpoints, mode, running state)
-        scomm: The sComm section of device_status (with hold type)
-        sfans: The sFanS section of device_status (with fan mode)
-    
-    Returns:
-        ClimateDevice model with dual mode (heating/cooling) support
-    
-    Raises:
-        KeyError: If required fields missing (caught as PARSING_EXCEPTIONS)
-    """
-    is_heating = ther["SystemMode"] == SystemMode.HEAT
-    fan_mode = sfans.get("FanMode", FanMode.AUTO)
-    hold_type = _hold_type(scomm, unique_id)
-    running_state = ther.get("RunningState", DEFAULT_RUNNING_STATE)
-
-    return ClimateDevice(
-        **_climate_common_args(device_status, unique_id),
-        current_humidity=None,
-        current_temperature=ther["LocalTemperature_x100"] / TEMPERATURE_SCALE,
-        target_temperature=(ther["HeatingSetpoint_x100"] / TEMPERATURE_SCALE)
-        if is_heating
-        else (ther["CoolingSetpoint_x100"] / TEMPERATURE_SCALE),
-        max_temp=(ther.get("MaxHeatSetpoint_x100", 4000) / TEMPERATURE_SCALE)
-        if is_heating
-        else (ther.get("MaxCoolSetpoint_x100", 4000) / TEMPERATURE_SCALE),
-        min_temp=(ther.get("MinHeatSetpoint_x100", 500) / TEMPERATURE_SCALE)
-        if is_heating
-        else (ther.get("MinCoolSetpoint_x100", 500) / TEMPERATURE_SCALE),
-        hvac_mode=HVAC_MODE_HEAT
-        if ther["SystemMode"] == SystemMode.HEAT
-        else HVAC_MODE_COOL
-        if ther["SystemMode"] == SystemMode.COOL
-        else HVAC_MODE_AUTO,
-        hvac_action=CURRENT_HVAC_OFF
-        if hold_type == HoldType.STANDBY
-        else CURRENT_HVAC_IDLE
-        if running_state == RunningState.IDLE
-        else CURRENT_HVAC_HEAT
-        if is_heating and running_state == RunningState.FAN_COIL_HEATING
-        else CURRENT_HVAC_HEAT_IDLE
-        if is_heating
-        else CURRENT_HVAC_COOL
-        if running_state == RunningState.FAN_COIL_COOLING
-        else CURRENT_HVAC_COOL_IDLE,
-        hvac_modes=[HVAC_MODE_HEAT, HVAC_MODE_COOL, HVAC_MODE_AUTO],
-        preset_mode=PRESET_OFF
-        if hold_type == HoldType.STANDBY
-        else PRESET_PERMANENT_HOLD
-        if hold_type == HoldType.PERMANENT_HOLD
-        else PRESET_ECO
-        if hold_type == HoldType.ECO
-        else PRESET_TEMPORARY_HOLD
-        if hold_type == HoldType.TEMPORARY_HOLD
-        else PRESET_FOLLOW_SCHEDULE,
-        preset_modes=[
-            PRESET_OFF,
-            PRESET_PERMANENT_HOLD,
-            PRESET_ECO,
-            PRESET_TEMPORARY_HOLD,
-            PRESET_FOLLOW_SCHEDULE,
-        ],
-        fan_mode=FAN_MODE_OFF
-        if fan_mode == FanMode.OFF
-        else FAN_MODE_HIGH
-        if fan_mode == FanMode.HIGH
-        else FAN_MODE_MEDIUM
-        if fan_mode == FanMode.MEDIUM
-        else FAN_MODE_LOW
-        if fan_mode == FanMode.LOW
-        else FAN_MODE_AUTO,
-        fan_modes=[
-            FAN_MODE_AUTO,
-            FAN_MODE_HIGH,
-            FAN_MODE_MEDIUM,
-            FAN_MODE_LOW,
-            FAN_MODE_OFF,
-        ],
-        locked=device_status.get("sTherUIS", {}).get("LockKey", 0) == 1,
-        supported_features=(
-            SUPPORT_TARGET_TEMPERATURE | SUPPORT_PRESET_MODE | SUPPORT_FAN_MODE
-        ),
-    )
-
-
-def _parse_climate_device(device_status: dict[str, Any]) -> ClimateDevice | None:
-    """Parse one climate device detail payload."""
-    unique_id = device_status.get("data", {}).get("UniID")
-    if unique_id is None:
-        return None
-
-    th = device_status.get("sIT600TH")
-    if isinstance(th, dict):
-        return _parse_it600th_climate_device(device_status, unique_id, th)
-
-    ther = device_status.get("sTherS")
-    scomm = device_status.get("sComm")
-    sfans = device_status.get("sFanS")
-    if isinstance(ther, dict) and isinstance(scomm, dict) and isinstance(sfans, dict):
-        return _parse_fan_coil_climate_device(
-            device_status,
-            unique_id,
-            ther,
-            scomm,
-            sfans,
-        )
-
-    return None
+    _walk(data)
+    return flattened
 
 
 class IT600Gateway:
@@ -1003,7 +536,7 @@ class IT600Gateway:
             devices,
             "cover",
             "_cover_devices",
-            _parse_cover_device,
+            parse_cover_device,
             self._send_cover_update_callback,
             send_callback,
         )
@@ -1017,7 +550,7 @@ class IT600Gateway:
             devices,
             "switch",
             "_switch_devices",
-            _parse_switch_device,
+            parse_switch_device,
             self._send_switch_update_callback,
             send_callback,
         )
@@ -1031,7 +564,7 @@ class IT600Gateway:
             devices,
             "sensor",
             "_sensor_devices",
-            _parse_sensor_device,
+            parse_sensor_device,
             self._send_sensor_update_callback,
             send_callback,
         )
@@ -1045,7 +578,7 @@ class IT600Gateway:
             devices,
             "binary sensor",
             "_binary_sensor_devices",
-            _parse_binary_sensor_device,
+            parse_binary_sensor_device,
             self._send_binary_sensor_update_callback,
             send_callback,
         )
@@ -1059,7 +592,7 @@ class IT600Gateway:
             devices,
             "climate",
             "_climate_devices",
-            _parse_climate_device,
+            parse_climate_device,
             self._send_climate_update_callback,
             send_callback,
         )
@@ -1172,6 +705,95 @@ class IT600Gateway:
 
         device_id = self._validate_device_id(device_id)
         return self._sensor_devices.get(device_id)
+
+    async def fetch_sq610_properties(
+        self,
+        device_ids: Sequence[str] | None = None,
+    ) -> dict[str, dict[str, Any]]:
+        """Fetch flattened raw gateway properties for cached SQ610 devices.
+
+        Args:
+            device_ids: Optional SQ610 climate device IDs. When omitted, all
+                cached SQ610 climate devices are fetched.
+
+        Raises:
+            KeyError: If a requested device is not in the climate cache.
+            ValueError: If a requested device is not an SQ610 model.
+            IT600ConnectionError: If the gateway cannot be reached.
+            IT600CommandError: If the gateway response is invalid or rejected.
+        """
+        if isinstance(device_ids, str):
+            raise TypeError("device_ids must be a sequence of strings, not a string")
+
+        if device_ids is None:
+            devices = [
+                device
+                for device in self._climate_devices.values()
+                if is_sq610_model(device.model)
+            ]
+        else:
+            devices = []
+            for device_id in device_ids:
+                device = self._require_device(
+                    device_id,
+                    self._climate_devices,
+                    "climate",
+                )
+                if not is_sq610_model(device.model):
+                    raise ValueError(f"climate device {device_id!r} is not an SQ610")
+                devices.append(device)
+
+        if not devices:
+            return {}
+
+        response = await self._make_encrypted_request(
+            "read",
+            {
+                "requestAttr": "deviceid",
+                "id": [{"data": device.data} for device in devices],
+            },
+        )
+
+        raw_properties: dict[str, dict[str, Any]] = {}
+        for device_status in _response_items(response, "SQ610 device detail"):
+            unique_id = device_status.get("data", {}).get("UniID")
+            if isinstance(unique_id, str):
+                raw_properties[unique_id] = _flatten_dict(device_status)
+
+        return raw_properties
+
+    async def write_sq610_property(
+        self,
+        device_id: str,
+        prop: str,
+        value: int,
+    ) -> None:
+        """Write one raw SQ610 `sIT600TH` property.
+
+        This intentionally exposes only the SQ610 write path currently required
+        by Home Assistant instead of making the full encrypted request API
+        public.
+        """
+        device = self._require_device(device_id, self._climate_devices, "climate")
+        if not is_sq610_model(device.model):
+            raise ValueError(f"climate device {device_id!r} is not an SQ610")
+
+        prop = _validate_non_empty_string(prop, "prop")
+        if isinstance(value, bool) or not isinstance(value, int):
+            raise TypeError("value must be an integer")
+
+        await self._make_encrypted_request(
+            "write",
+            {
+                "requestAttr": "write",
+                "id": [
+                    {
+                        "data": device.data,
+                        "sIT600TH": {prop: int(value)},
+                    }
+                ],
+            },
+        )
 
     async def set_cover_position(self, device_id: str, position: int) -> None:
         """Move a cover to a position where 0 is closed and 100 is open."""
